feat: add required attribute to cron.xml
In this commit, we have improved the handling of cron tasks and introduced the 'required'
attribute. Changes made include:

- Added 'required' attribute to three cron exec in `cron.xml` to enforce their creation.
- Updated `EventHandler` to only create `required` crons, and to stop creating non-required crons
when the `$onlyRequired` argument is set to `true`.
- Improved how crons are created with different scopes by replacing the switch statement with a
`match` expression in `EventHandler`.
- Added the handling of the 'required' attribute in `Manager` to ensure all cron objects are aware
of if they are required or not. This attribute is set to `true` for required crons and `false` if
not specified or not required.
- Simplified the array replacement in `EventHandler`.

This improvement in how cron tasks are managed will result in more efficient and robust operation
of the system.